organized and managed all interfaces
