+++ src/pages/guest/AdminDemo.tsx
@@ -1,9 +1,9 @@
 import React, {useEffect, useState} from "react"
 import Nav from "../../components/Nav"
-import {Input, handleLogin, GetUsers, User} from "../../hooks/LoginHooks"
+import {Input, handleLogin, GetUsers} from "../../hooks/LoginHooks"
 import {ButtonSubmit} from "../../components/Button"
 import Footer from "../../components/Footer"
-
+import {User} from "../../middleware/Interfaces"
 export default function AdminDemo(){
 
     const name = "Bob The Builder"
